Fix type mismatch in HttpMcpConnection timeout configuration
OkHttp's connectTimeout and readTimeout methods expect Long values,
not Int. Removed unnecessary .toInt() conversions.

+++ shared/src/jvmMain/kotlin/link/socket/ampere/agents/tools/mcp/connection/HttpMcpConnection.jvm.kt
@@ -27,8 +27,8 @@ actual class HttpClientHandler {
                 httpClient = HttpClient(OkHttp) {
                     engine {
                         config {
-                            connectTimeout(timeoutMs.toInt(), java.util.concurrent.TimeUnit.MILLISECONDS)
-                            readTimeout(timeoutMs.toInt(), java.util.concurrent.TimeUnit.MILLISECONDS)
+                            connectTimeout(timeoutMs, java.util.concurrent.TimeUnit.MILLISECONDS)
+                            readTimeout(timeoutMs, java.util.concurrent.TimeUnit.MILLISECONDS)
                         }
                     }
                 }
